import-rapport-odoo/requirements.txt
python-dotenv>=1.0.0
colorama>=0.4.6
tqdm>=4.65.0
